pyproject.toml
[tool.poetry]
name = "{project_name}"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
pytz = "^2021.1"
celery = "^5.1.2"
Pillow = "^8.3.2"
gunicorn = "^20.1.0"
gevent = "^21.8.0"
djangorestframework = "^3.12.4"
django-auth-framework = "^1.0.3"
django-oauth-toolkit = "^1.5.0"
django-cors-headers = "^3.8.0"
django-redis-cache = "^3.0.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
